+++ source/renderer/app/Routes.js
@@ -22,7 +22,6 @@ import StakingRewardsPage from './containers/staking/StakingRewardsPage';
 import StakePoolsListPage from './containers/staking/StakePoolsListPage';
 import StakingCountdownPage from './containers/staking/StakingCountdownPage';
 import RedeemItnRewardsContainer from './containers/staking/RedeemItnRewardsContainer';
-import HardwareWallet from './containers/hardware-wallet/HardwareWallet';
 import Wallet from './containers/wallet/Wallet';
 import WalletAddPage from './containers/wallet/WalletAddPage';
 import WalletSummaryPage from './containers/wallet/WalletSummaryPage';
@@ -107,14 +106,6 @@ export const Routes = withRouter(() => (
           path={ROUTES.PAPER_WALLET_CREATE_CERTIFICATE}
           component={PaperWalletCreateCertificatePage}
         />
-        <Route path={ROUTES.HARDWARE_WALLETS.ROOT}>
-          <HardwareWallet>
-            <Route
-              path={ROUTES.HARDWARE_WALLETS.SUMMARY}
-              component={WalletSummaryPage}
-            />
-          </HardwareWallet>
-        </Route>
         <Route path={ROUTES.STAKING.ROOT}>
           <Staking>
             <Route