+++ source/renderer/app/containers/wallet/dialogs/WalletSendConfirmationDialogContainer.js
@@ -3,6 +3,7 @@ import React, { Component } from 'react';
 import { inject, observer } from 'mobx-react';
 import type { StoresMap } from '../../../stores/index';
 import type { ActionsMap } from '../../../actions/index';
+import type { HwDeviceStatus } from '../../../domains/Wallet';
 import WalletSendConfirmationDialog from '../../../components/wallet/WalletSendConfirmationDialog';
 
 type Props = {
@@ -15,6 +16,8 @@ type Props = {
   amountToNaturalUnits: (amountWithFractions: string) => string,
   currencyUnit: string,
   onExternalLinkClick: Function,
+  hwDeviceStatus: HwDeviceStatus,
+  isHardwareWallet: boolean,
 };
 
 @inject('actions', 'stores')
@@ -23,7 +26,20 @@ export default class WalletSendConfirmationDialogContainer extends Component<Pro
   static defaultProps = { actions: null, stores: null };
 
   handleWalletSendFormSubmit = (values: Object) => {
-    this.props.actions.wallets.sendMoney.trigger(values);
+    if (values.isHardwareWallet) {
+      this.props.actions.hardwareWallets.sendMoney.trigger({});
+    } else {
+      this.props.actions.wallets.sendMoney.trigger(values);
+    }
+  };
+
+  handleInitiateTransaction = () => {
+    const { stores } = this.props;
+    const { wallets, hardwareWallets } = stores;
+    const { active: activeWallet } = wallets;
+    if (!activeWallet)
+      throw new Error('Active wallet required for WalletSendPage.');
+    hardwareWallets.initiateTransaction({ walletId: activeWallet.id });
   };
 
   render() {
@@ -36,13 +52,29 @@ export default class WalletSendConfirmationDialogContainer extends Component<Pro
       transactionFee,
       amountToNaturalUnits,
       currencyUnit,
+      hwDeviceStatus,
+      isHardwareWallet,
     } = this.props;
     const { stores } = this.props;
     const { sendMoneyRequest, active: activeWallet } = stores.wallets;
+    const {
+      _resetTransaction: resetHardwareWalletTransaction,
+      sendMoneyRequest: sendMoneyExternalRequest,
+      isTransactionPending,
+    } = stores.hardwareWallets;
     const { isFlight } = global;
 
     if (!activeWallet)
       throw new Error('Active wallet required for WalletSendPage.');
+
+    const isSubmitting =
+      (!isHardwareWallet && sendMoneyRequest.isExecuting) ||
+      (isHardwareWallet &&
+        (sendMoneyExternalRequest.isExecuting || isTransactionPending));
+
+    const error = isHardwareWallet
+      ? sendMoneyExternalRequest.error
+      : sendMoneyRequest.error;
 
     return (
       <WalletSendConfirmationDialog
@@ -52,15 +84,20 @@ export default class WalletSendConfirmationDialogContainer extends Component<Pro
         transactionFee={transactionFee}
         amountToNaturalUnits={amountToNaturalUnits}
         onSubmit={this.handleWalletSendFormSubmit}
-        isSubmitting={sendMoneyRequest.isExecuting}
+        isSubmitting={isSubmitting}
         isFlight={isFlight}
         onCancel={() => {
           actions.dialogs.closeActiveDialog.trigger();
           sendMoneyRequest.reset();
+          resetHardwareWalletTransaction({ cancelDeviceAction: true });
         }}
-        error={sendMoneyRequest.error}
+        error={error}
         currencyUnit={currencyUnit}
         onExternalLinkClick={onExternalLinkClick}
+        hwDeviceStatus={hwDeviceStatus}
+        isHardwareWallet={isHardwareWallet}
+        onInitiateTransaction={this.handleInitiateTransaction}
+        walletName={activeWallet.name}
       />
     );
   }